Change remap filter behavior to match ip_allow.yaml (#9631) (#548)
This could possibly be considered a bug, but certainly it's
not a great behavior as it was. Before this patch, filters
(as defined and used with .definefilter and .activatefilter),
are always all evaluated, with a priority towards deny.

ip_allow.yaml however works by evaluating only filters up until
one matches, and whatever action has triggered, is used.

This patch changes these remap filters to follow the behavior
of ip_allow.yaml. First filter that matches is applied, and
evaluation of filters stops.

I would really have liked to see this in 9.2.x, but I'm not
so sure that it's a good idea since it does also break
compatibility even if we consider this a bug.

(cherry picked from commit 1cccda1b6d9d8bd31854dee986a879a285079dfe)

Co-authored-by: Leif Hedstrom <zwoop@apache.org>