don't define _pr_id rpm macro if it is empty string (#571)
* don't define _pr_id rpm macro if it is empty string

If $INTERNAL_GITHUB_PR is undefined (empty string) rpmbuild command line
args define the rpm macro _pr_id with an empty body. This  produces an
error that is ignored on el7, but not in el9. Only define the macro if
it is not null.

You can see this behavior in the build logs
     + rpmbuild -ba build/bazinga-trafficserver.spec --define '_version
       9.2.1' --define '_release 90.14314' --define '_archive
       bazinga-trafficserver-9.2.1.tar.gz' --define '_commit
       0d508efda92257b8bde7006f9af0107733eb72b8' --define '_gittag
       bazinga/9.2.1/90.14314' --define '_topdir
       /workspace/build/rpmbuild' --define '_pr_id ' --with llvm
     error: Macro %_pr_id has empty body
     error: Macro %_pr_id has empty body

* bash variable expansion strikes again

* fix build.sh, change specfile reqs for rhel9